Split up generic `RemoveSerializer`
Into three distinct ones for each case.
