+++ app/serializers/activitypub/remove_note_serializer.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+
+class ActivityPub::RemoveNoteSerializer < ActivityPub::Serializer
+  attributes :type, :actor, :target
+  has_one :proper_object, key: :object
+
+  def type
+    'Remove'
+  end
+
+  def actor
+    ActivityPub::TagManager.instance.uri_for(object.account)
+  end
+
+  def proper_object
+    ActivityPub::TagManager.instance.uri_for(object)
+  end
+
+  def target
+    ActivityPub::TagManager.instance.collection_uri_for(object.account, :featured)
+  end
+end
